Tidy code
