+++ .gitignore
@@ -1,2 +1,4 @@
 target/
+tmp/
+!tmp/.gitkeep
 Cargo.lock
