+++ judge-core/src/lib.rs
@@ -2,7 +2,7 @@ pub mod builder;
 pub mod compiler;
 pub mod error;
 pub mod executor;
-pub mod monitor;
+pub mod judge;
 pub mod result;
 pub mod sandbox;
 pub mod utils;
