--- tmp/icpc/data/secret/0.ans
@@ -1,1 +1,0 @@
-Hello! world!
