--- tmp/icpc/data/secret/0.in
@@ -1,1 +1,0 @@
-world!
