--- tmp/interactor
@@ -1,2 +1,0 @@
-world!
-Hello! world!
